refactor(reporter): Log and metrics reporting
Closes gravitee-io/issues#836
