+++ src/main/java/io/gravitee/reporter/api/common/Response.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.reporter.api.http;
+package io.gravitee.reporter.api.common;
 
 import io.gravitee.common.http.HttpHeaders;
