+++ src/main/java/io/gravitee/reporter/api/health/EndpointStatus.java
@@ -17,8 +17,8 @@ package io.gravitee.reporter.api.health;
 
 import io.gravitee.common.utils.UUID;
 import io.gravitee.reporter.api.AbstractMetrics;
-import io.gravitee.reporter.api.http.Request;
-import io.gravitee.reporter.api.http.Response;
+import io.gravitee.reporter.api.common.Request;
+import io.gravitee.reporter.api.common.Response;
 
 import java.util.ArrayList;
 import java.util.List;
